correct schemeshard usage for index operations
