+++ ydb/core/protos/flat_scheme_op.proto
@@ -421,6 +421,25 @@ message TOlapColumnDescription {
     optional TDictionaryEncodingSettings DictionaryEncoding = 9;
 }
 
+message TBloomFilter {
+    optional double FalsePositiveProbability = 1 [default = 0.1];
+    optional uint64 MaxBytesCount = 2 [default = 8196];
+}
+
+message TOlapIndexDescription {
+    // This id is auto-generated by schemeshard
+    optional uint32 Id = 1;
+
+    // Column name is required, some column names are reserved
+    optional string Name = 2;
+    optional string ClassName = 3;
+    optional TCompressionOptions Compression = 4;
+
+    oneof IndexImplementation {
+        TBloomFilter BloomFilter = 40;
+    }
+}
+
 enum EColumnTableEngine {
     COLUMN_ENGINE_NONE = 0;
     COLUMN_ENGINE_REPLACING_TIMESERIES = 1;
@@ -456,6 +475,7 @@ message TColumnTableSchema {
     optional TCompressionOptions DefaultCompression = 8;
 
     optional bool CompositeMarks = 9 [ default = false ];
+    repeated TOlapIndexDescription Indexes = 10;
 }
 
 message TAlterColumnTableSchema {
@@ -463,6 +483,8 @@ message TAlterColumnTableSchema {
     //optional TCompressionOptions DefaultCompression = 5;
     repeated TOlapColumnDescription DropColumns = 6;
     repeated TOlapColumnDiff AlterColumns = 7;
+    repeated TOlapIndexDescription UpsertIndexes = 8;
+    repeated string DropIndexes = 9;
 }
 
 // Schema presets are used to manage multiple tables with the same schema
