+++ ydb/core/tx/columnshard/engines/scheme/indexes/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    abstract.cpp
+)
+
+PEERDIR(
+    ydb/core/protos
+    ydb/core/formats/arrow
+)
+
+END()
