+++ ydb/core/tx/schemeshard/olap/common/common.cpp
@@ -1,0 +1,4 @@
+#include "common.h"
+
+namespace NKikimr::NSchemeShard {
+}
