+++ ydb/core/tx/schemeshard/olap/common/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    common.cpp
+)
+
+PEERDIR(
+    ydb/library/ydb_issue
+    ydb/core/base
+)
+
+END()
