+++ ydb/core/tx/schemeshard/olap/indexes/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    schema.cpp
+    update.cpp
+)
+
+PEERDIR(
+    ydb/services/bg_tasks/abstract
+)
+
+END()
