+++ ydb/core/tx/schemeshard/olap/schema/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    schema.cpp
+    update.cpp
+)
+
+PEERDIR(
+    ydb/core/tx/schemeshard/olap/columns
+    ydb/core/tx/schemeshard/olap/indexes
+)
+
+END()
