+++ ydb/core/tx/schemeshard/olap/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+PEERDIR(
+    ydb/core/tx/schemeshard/olap/columns
+    ydb/core/tx/schemeshard/olap/indexes
+    ydb/core/tx/schemeshard/olap/schema
+    ydb/core/tx/schemeshard/olap/common
+    ydb/core/tx/schemeshard/olap/operations
+)
+
+END()
